+++ src/SophiaWin11.Core/Abstractions/IThemeService.cs
@@ -1,0 +1,8 @@
+namespace SophiaWin11.Core.Abstractions;
+
+public interface IThemeService
+{
+    bool IsThemeApplied { get; }
+
+    void ApplyTheme();
+}